+++ primus-common/src/main/java/com/bytedance/primus/api/protocolrecords/RegisterRequest.java
