+++ primus-common/src/main/java/com/bytedance/primus/api/records/ClusterSpec.java
