+++ primus-common/src/main/java/com/bytedance/primus/api/records/ExecutorSpec.java
