+++ primus-common/src/main/java/com/bytedance/primus/api/records/ExecutorState.java
