+++ primus-common/src/main/java/com/bytedance/primus/api/records/TaskType.java
