+++ primus-common/src/main/java/com/bytedance/primus/api/wrapper/RegisterRequestProtoWrapper.java
