+++ primus-common/src/main/java/com/bytedance/primus/api/wrapper/RegisterResponseProtoWrapper.java
